Extracted "Tab" and "TabTitleConverter" classes to the separate file.
